EX4_PublicAPI/requirements.txt
Flask==3.0.2
Pillow==10.3.0
numpy==1.24.1
tensorflow==2.15.0
